shard mapper tests
